check sub-permissions
